rename package name from `generator` to `tweakers`
