+++ common/src/main/java/org/astraea/common/balancer/algorithms/GreedyBalancer.java
@@ -28,7 +28,7 @@ import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
 import org.astraea.common.admin.Replica;
 import org.astraea.common.balancer.Balancer;
-import org.astraea.common.balancer.generator.ShuffleTweaker;
+import org.astraea.common.balancer.tweakers.ShuffleTweaker;
 import org.astraea.common.balancer.log.ClusterLogAllocation;
 import org.astraea.common.cost.ClusterCost;
